Add a typed, stampede-aware cache type
This adds a type Cache[T] which can be used as a typed interface to a
distributed cache stored in Redis.

The public interface of the cache is a single function, Get, which takes
both a key and a function which can refresh the cache or fetch a value
in the case of a cache miss.

The cache can be configured with two separate lifetimes:

1. The "fresh" lifetime, which is the duration for which the data can be
   considered fresh. Once the data is no longer fresh, a background
   cache update will be triggered.

2. The "stale" lifetime, which is the duration after which the data will
   expire from cache altogether. Until the "stale" lifetime is up, the
   cache will return stale data if necessary (e.g. while waiting for a
   background cache update to complete).

Background cache updates are coordinated using a distributed lock (also
stored in Redis) using the "lock" package to avoid cache stampedes when
the data in the cache becomes stale.