Update LaunchDarkly SDK to v6
This release adds contexts, which we want for feature flagging based on
versions.

Quite a few things moved about in this new version, so this change is a
bit involved! I'll leave adding methods for flagging based on versions
to a follow up commit, to keep this as small as possible.