Add WithFlagContext and FlagContextFromContext
These methods can be used to store a LaunchDarkly context object on a Go
context and later retrieve it from that Go context or its children. This
makes it possible to construct per-request flag contexts in one place
and then pass them through the call tree on the Go context object.