Fall back to fetching from upstream on Redis errors
The intent of the cache is to reduce the load going upstream and
partially decouple the availability of the service using the cache from
the availability of the upstream service.

As far as possible we want to avoid the possibility that caching
introduces a new source of failure to the system, and so we need to
tolerate the possibility that Redis is unavailable (during failover
events or general network weather).

This commit changes the error behaviour of the cache: if we encounter an
unexpected error communicating with Redis we fall back to calling the
passed fetcher and loading data from source.

This *does* introduce new risks. Failure of a heavily-used cache under
this new logic will trigger huge increases in load going upstream,
potentially resulting in cascading effects. This may be something we
need to address in future with appropriate circuit-breakers.

For now we want the availability of the data accessed through the cache
to be the *sum* of the availability of data in cache and that of the
upstream service (and not the product of those availabilities) so this
is a better behaviour.