Switch precedence of state and arguments
The way we're using this in API has ended up being a little different
from what I had originally envisioned. We now always know the request
identity (and hence any custom limits) before checking the rate limiter,
so the rate and capacity provided as arguments should override anything
set in the bucket.

This does mean that there's no method in the Go library which can query
the bucket without overriding rate/capacity, but we can always add that
later if it becomes necessary.