Fix a bug where we sometimes didn't set a TTL
I found a bunch of rate limiter keys in production that didn't have a
TTL set on them, and it turns out to be the stupidest bug I've written
in a while. We were calling EXPIRE before the key even existed. If a
given key only checks a rate limit exactly once, the key won't have a
TTL set on it.

The fix is to swap the order of two lines of code. Also added a
regression test to make sure it works.