Add a TraceOptionsProcessor which automatically annotates spans
This adds a SpanProcessor, TraceOptionsProcessor, which automatically
adds the "meta.replicate.disable_sampling" span attribute whenever the
SampleMode trace option is set to "always".

Together with the previous commit, this means that any Go application in
our stack will automatically add this attribute to its spans when the
tracestate contains our trace options keys. One caveat is that this only
applies for spans started after the tracestate has been added to the
span context.