Don't attempt to replace the current SpanContext
In general, it turns out that replacing the current SpanContext is not a
good idea. Indeed, the OTel documentation [says][1]:

> Please note, since `SpanContext` is immutable, it is not possible to
> update `SpanContext` with a new `TraceState`. Such changes then make
> sense only right before `SpanContext` propagation or telemetry data
> exporting. In both cases, `Propagator`s and `SpanExporter`s may create
> a modified `TraceState` copy before serializing it to the wire.

So this commit updates how we set `TraceOptions` *during* a span.
Namely, when we update `TraceOptions` using `WithTraceOptions`, we no
longer modify the `SpanContext`. Instead, the value of `TraceOptions` is
stored on the `Context`, from where it can be retrieved by
`TraceOptionsPropagator` at propagation time.

This means that at propagation we will respect any `TraceOptions` set in
the `TraceState`, *unless* TraceOptions have been set in the `Context`,
in which case those will be used instead.

[1]: https://opentelemetry.io/docs/specs/otel/trace/api/#tracestate