Add FlagBlueGreen to conventionalise blue-green rollouts
This exposes what is essentially a boolean flag with a default, but with
names and conventions about how boolean values map to those names so
that it can work the same way everywhere we use it.

In code this would look like:

    cogRolloutDefault := flags.ResultGreen

    if flags.FlagBlueGreen(&flagContext, "cog-rollout", cogRolloutDefault) == flags.ResultBlue {
      cogVersion = cogVersionBlue
    } else {
      cogVersion = cogVersionGreen
    }

Once the rollout is complete and stable in the "blue" state, the
operator would update the value of cogRolloutDefault to be
`flags.ResultBlue`.