Add DroppedDataProcessor to warn about dropped data
By default, attributes/events/links are dropped silently when span
limits are exceeded. This adds a SpanProcessor which will log a warning
when this occurs instead.

There is a risk that this will create performance issues when data
starts to be dropped, but implementing throttling right now seems like
overkill.