Steps toward an RFC 9241 HTTP Message Signatures implementation
We want to use request signing to authenticate service-to-service
traffic within Replicate. Request signing is an attractive option for a
number of reasons. Two important ones:

1. we authenticate individual requests, not a communication channel
   shared between many requests (looking at you, mTLS)
2. we have access to authentication data, signature parameters, etc., at
   the HTTP layer, which makes enforcing per-endpoint requirements much
   easier

This commit starts to lay the groundwork for an implementation of HTTP
Message Signatures in compliance with RFC 9241. This is by no means a
complete implementation of the spec, but it should already cover almost
everything needed for deployment at Replicate.

Notably, there is currently no support for signing responses, only
requests.

Currently only signing is implemented. Verification code will initially
only be needed in Python, although we'll likely want to add it here so
we can more effectively test this.