Remove tools.go
With more recent versions of Go, it's possible to specify a version when
using `go run`, which means we no longer need auxiliary tools to be
specific in `tools.go` and have their deps managed by `go.mod`.

This dramatically simplifies the dependency tree for the module when
built for production.