Simplify use of telemetry package
We almost never want to do anything other than log a warning if
telemetry isn't configured properly, so this commit removes the Start
function from the telemetry package, and instead just configures
telemetry on package initialization.

If OTEL_EXPORTER_OTLP_ENDPOINT is not set, we log a warning and don't
configure a tracer provider. This improves the experience in
development, where previously the batch span exporter would log an error
every few seconds because it was unable to export to the default
https://localhost:4318/v1/traces endpoint.