Add shuffle-shard queue write support
Implements one half of a shuffle-sharded queue implementation.
Specifically: the write half.

This wraps up the logic for writing to a shuffle-sharded queue so that
clients can simply provide the relevant parameters for the queue, and
the details of which stream is written to are hidden.

We calculate the possible streams (based on a shard key) in Go, and then
pass those into a Lua script in Redis which does the following:

1. Checks to see if it can update the queue "meta" key to reflect the
   current number of total streams.

   If the new value is greater than the current one it can do this
   immediately. If the new value is less than the current one it will
   check the lengths of the streams beyond the current set, and only
   update the meta key if all those streams are empty.

   For example: if we were previously writing to 16 streams (with
   indexes 0 to 15) and new requests arrive saying that we're writing to
   8 streams (0 to 7), then we'll check that streams 9 to 15 are empty
   before updating the meta key.

2. Finds the shortest stream out of those provided, or the first empty
   stream.

3. XADDs the provided message to the selected stream.

4. XADDs a short message to a `notifications` stream, which can be used
   to await new work when queues are not busy.

4. Sets a timeout/expiry on all the keys of the queue.

5. Returns the enqueued message ID.