Implement read support for sharded queues
This implements the read side of the sharded queue implementation. This
is quite fiddly, primarily because we cannot block/wait on a write to
one of N queues easily while also guaranteeing that we never pick up
more than one message at a time.

Infuriatingly, when the queue is "empty", then

  XREADGROUP ... COUNT 1 BLOCK 1000 STREAMS key1 key2 [...] > > [...]

will do the right thing. It will return immediately after the first XADD
on any of the monitored streams. The problem is that if there are items
immediately available in more than one of the specified streams, one per
stream will be claimed and returned.

This implementation works around this by adding yet another stream -- a
"notifications" stream -- the sole purpose of which is to accelerate the
next read by the client when the stream is mostly empty.

Every time we write to the queue we also put a tiny message on the
notifications stream, which can then be used by clients to block on
activity on the queue as a whole.

An individual blocking read, then, has three steps:

1. a non-blocking pass through all the queues
2. if nothing is found, block on the notifications queue
3. if we get a message, do another non-blocking pass through all the
   queues

This is inherently racy, but should hopefully work well enough for our
purposes and doesn't involve hammering Redis to scan for messages
repeatedly.

The read code also handles a number of other aspects of this problem:

- for new queues, the creation of the requested consumer group is
  handled automatically
- we will automatically read from the "default" stream if it exists and
  has items in it -- this will allow us to migrate everything to sharded
  queues without downtime
- the consumer group provided is used when reading from the new streams,
  but when reading from the default stream we use the stream name as a
  consumer group name -- this will allow us to migrate away from having
  different consumer group names for each queue