Introduce test helpers and run integration tests separately
This introduces a `test` package which contains some test helpers.

It also introduces a convention whereby tests with "Integration" in the
name are treated as integration tests and run separately from the rest
of the test suite using `-skip` and `-run` arguments to `go test` as
needed.

Integration tests that access an external shared resource such as Redis
often need to be run serially, so providing a common pattern for them
them will allow us to run non-integration tests as fast as possible.