Use a coordinated offset rather than a "random" one
Dom pointed out that using a random (or, as in our case, "random")
offset for choosing where to start reading doesn't actually result in
fair reads.

Consider the contrived but quite possible case of a 64-stream queue in
which tenant A is allocated streams [0,4), tenant B is allocated streams
[4,8), and no other tenants are using the queue.

Any read which starts at offsets in the range [4,8) will fetch a message
for tenant B, but a read starting at any other offset (all 60 of them)
will fetch a message for tenant A.

This updates the read code to use a globally coordinated offset stored
in the "meta" key to ensure fairness across reads, still without
requiring consumers to maintain state. Any time a message is found, the
offset is updated to point to the *next* stream in the queue, and then
reads start at the recorded offset. This should ensure that the first
read round-robins through all queues that have messages in them.