[lock] support multiple redis backends
We want to migrate from one redis instance to another.  To support this, we
now allow acquiring locks on multiple redis instances simultaneously.

Locks will be acquired in order (1,2,3) and released in reverse order (3,2,1).

If we fail to acquire a lock on a given instance, we go back and release all
locks acquired so far.