[cache,lock] support multiple redis backends (#133)
* [lock] support multiple redis backends

We want to migrate from one redis instance to another.  To support this, we
now allow acquiring locks on multiple redis instances simultaneously.

Locks will be acquired in order (1,2,3) and released in reverse order (3,2,1).

If we fail to acquire a lock on a given instance, we go back and release all
locks acquired so far.

* [cache] Support multiple redises

We want to migrate from one redis instance to another.  To support this, we now
allow cache to use multiple redis instances simultaneously.

Fetch operations will stop at the first cache hit.  If none hit, we return the
result from the last cache - whether negative cache, soft miss, or hard miss.
We don't try to be clever about this - after all, this is supposed to be a
short-term thing.

Set operations will acquire a pessimistic lock to ensure we don't get into an
inconsistent state between the multiple instances.  Sets are then written to all
backends.  If one backend fails we still attempt to write to all the others.

* PR feedback