Remove unused pickup delay telemetry
This is being added to a span that is unlikely to be recorded by most
clients (because they will typically only want to record spans when they
get a queue message, not for every attempt to fetch a queue message).

We've added the same telemetry in a more useful place in
replicate/director#291.