Add cancelation metadata to Redis
Add some cancelation metadata to redis that allows API to introspect the
stream and message ID from the prediction ID. This allows api to perform
an `XDEL` on the redis message when a cancelation is received. The only
time an `XDEL` should happen is if the prediction has not started.

API will be responsible to check `XPENDING` and/or the prediciton cache.
Additionally API should `DEL` the meta key when it performs a
cancelation or when the prediction completes.

key format: `meta:cancelation:<prediction_id>` and it contains json
in the form of: `{"stream_id": "<stream key>", "msg_id": "<redis message
id>"}`