[queue] Fix off-by-one error
Our loop bounds were wrong here.

This isn't a big deal in len.lua because XLEN on a nonexistent key returns 0.

It's not a big deal in read.lua because we immediately take (offset + idx) %
streams, folding into the range [0, streams) anyway.  It just means we check the
original stream twice in the loop.

However this breaks if you copy-paste the pattern for some other purpose, like
calling XINFO GROUPS on a key.  That command errors if the key doesn't exist.