[queue] remove legacy (#158)
* [queue] Fix off-by-one error

Our loop bounds were wrong here.

This isn't a big deal in len.lua because XLEN on a nonexistent key returns 0.

It's not a big deal in read.lua because we immediately take (offset + idx) %
streams, folding into the range [0, streams) anyway.  It just means we check the
original stream twice in the loop.

However this breaks if you copy-paste the pattern for some other purpose, like
calling XINFO GROUPS on a key.  That command errors if the key doesn't exist.

* [queue] Remove legacy code for checking old base streams

This is now fully rolled out, we have no base streams any more.  I have checked
this in all our clusters with:

    KEYS input:prediction:??-????????????????????????????????

and found a bunch of empty streams with no TTLs.  I manually deleted them.