[queue] create PendingCount (to replace Lag)
This adds a new PendingCount command which returns the aggregate number of
pending messages in a queue.

-----

I tried working with Lag but it has the problem that it is fundamentally
heuristic, based on keeping a track of two counters:

- how many messages have ever been added to the stream
- how many messages have ever been read by the given consumer group

"Lag" is then the difference between these counters - the number of messages
that have been added to the stream but not read by the consumer group.

Unfortunately, these counters can desync horribly if you ever XDEL a message
before it gets read by a consumer group.  This is because deleting a message
does not decrement the "added to stream" counter, nor does it increment the
"read by consumer group" counter.

Currently we rely on being able to XDEL messages.  Fixing that would be
nontrivial.

But we can measure the size of the PEL with XPENDING, and we can measure the
length of the stream with XLEN, so we can calculate lag as PendingCount() -
Len().