[queue] create Stats (to replace PendingCount)
This is just like #160, but we calculate length and pending count atomically.

Before this change, I was subtracting the pending count from the length to get a
"waiting messages" metric; but as these numbers could not be measured atomically
I sometimes saw negative waiting messages.

This changes PendingCount() into Stats() which returns both length and pending
count.