Update our len() function to account for pending
To ensure that our len function returns data that is more accurate to
calculate the number of instances we need, subtract the XPENDING count
from the XLEN count. This reduces the cases where we a new version or
scaling config changes will result in a complete duplication of the pods
simply because pending elements in the queue are counted as being
in-queue.

Use of XPENDING will allow us to subtract anything that has been claimed
from the metric calculation.