Support rates less than 1 for rate limiting (#198)
Support a floating point rate. In the implementation as current, if multiple calls are made before a whole token is filled, we will starve the process by continually updating the last_fill_time whether we filled a token or not, meaning for example if you filled one token every 10s but called the rate limiter every 1s you would never get a token. Removing this bug is necessary for lower fill rates because it is very unlikely to happen when you have 50 rps. We also need to update the time_to_fill_bucket because we can no longer assume that the number of tokens needed to fill is a good enough estimate since it may be multiple seconds to fill a single token, so for example a bucket with a capacity of 1 if you call 1 per second, there's no guarantee you've called it in the first second or the 9th second. Therefore we need to return the delta from the last filled token.